feat(*): update data exports and rollup config
